Add scalers (center standardize, minmax, maxabs)
An implementation for normalize was also made, but is currently commented out.
A statistics function `item_stats` was added to compute min, max, mean, and std. for each elements in an item.
Some restructure had to be made for _optional_argument_indexed_transform and how arguments are passed.
They are now wrapped in a list, in order to accommodate multiple arguments.
Readme was updated as well.